Get initial outgoing teleport support to actually work.
This works some of the time (though occasionally the user gets logged off instead), but isn't sending the destination
position or a bunch of other, hopefully fairly non-critical, stuff to the target sim.

Also, we now parse the wearables on an incoming teleport, or the avatar's appearance gets screwed up on teleporting in
and out again.