More teleportation work. Still not quite right.
This tracks the seed capabilities for other regions that we're given, but it doesn't send them onto the next region
(and doesn't add our seed cap to the list in the login case).

Also, it looks like the problem with getting spat back to the original sim may be a viewer or OpenSim issue with
teleporting out of a region and back in in rapid succession. It's quite hard to pin down, though.