Delay agent deletion on teleport out.
This is needed because we get notified when the destination receives the
CompleteAgentMovement message, yet the client doesn't fully change sims
until it's got a seed cap response from the destination too. If we
delete the agent immediately, the user ends up unexpectedly getting
logged out on teleports (due to the dead EventQueueGet, probably).