We now check for and prevent overflows of the script stack.
Note that, since the check is only done on function entry, it triggers before the program actually reaches the end of
the stack, and indeed on code that may never actually cause a stack overflow. It's unlikely that this will be an issue,
though.

Also, we now send script error messages back to the viewer as appropriate, mostly.