Fix unwinding after an unimplemented opcode so it doesn't segfault.
The problem is that, if the script tried to execute an unimplemented opcode, IP was left pointing at the next
instruction, so the stack unwinding code assumed that the instruction had been executed when it hadn't. This caused a
moderately impressive segfault. We fix the issue by decrementing IP so that it points at the unimplemented instruction.